agentui/requirements.txt
fastapi
uvicorn
python-dotenv
azure-ai-projects
azure-identity
azure-core